Added default values for Yummly terms settings
This commit closes #347. No default value had been set for the Yummly
settings, and as such, it was failing to find a default and was instead
retaining a persistent default from a previous option that the system
had already cycled through and pulled a value from. By adding the
defaults, it should now always appear as blank fields until users add
something to them.